feat: split renderer base
