+++ core/include/karin/graphics/offscreen_renderer.h
@@ -1,0 +1,12 @@
+#ifndef CORE_KARIN_GRAPHICS_OFFSCREEN_RENDERER_H
+#define CORE_KARIN_GRAPHICS_OFFSCREEN_RENDERER_H
+
+namespace karin
+{
+class OffscreenRenderer
+{
+public:
+};
+} // karin
+
+#endif //CORE_KARIN_GRAPHICS_OFFSCREEN_RENDERER_H
