+++ core/src/graphics/offscreen_renderer.cpp
@@ -1,0 +1,1 @@
+#include <karin/graphics/offscreen_renderer.h>
