+++ core/src/graphics/window_renderer.cpp
@@ -3,11 +3,7 @@
 #include "platform.h"
 #include "renderer_impl.h"
 
-#include <stb_image/stb_image.h>
-
 #include <vector>
-#include <stdexcept>
-#include <cstring>
 
 namespace karin
 {
@@ -18,11 +14,6 @@ WindowRenderer::WindowRenderer(Window* window)
 }
 
 WindowRenderer::~WindowRenderer() = default;
-
-void WindowRenderer::addDrawCommand(std::function < void(GraphicsContext &) > command)
-{
-    m_drawCommands.push_back(std::move(command));
-}
 
 void WindowRenderer::update() const
 {
@@ -74,35 +65,5 @@ void WindowRenderer::update() const
             m_impl->finishResizing();
         }
     );
-}
-
-void WindowRenderer::cleanUp()
-{
-    m_impl->cleanUp();
-}
-
-void WindowRenderer::setClearColor(const Color& color)
-{
-    m_impl->setClearColor(color);
-}
-
-Image WindowRenderer::createImage(const std::string& filePath)
-{
-    int width, height, channels;
-    stbi_uc* data = stbi_load(filePath.c_str(), &width, &height, &channels, 4);
-    if (!data)
-    {
-        throw std::runtime_error("Failed to load image: " + filePath);
-    }
-
-    std::vector<std::byte> imageData(width * height * 4);
-    memcpy(imageData.data(), data, imageData.size());
-
-    return m_impl->createImage(imageData, static_cast<uint32_t>(width), static_cast<uint32_t>(height));
-}
-
-Image WindowRenderer::createImage(const std::vector<std::byte>& data, uint32_t width, uint32_t height)
-{
-    return m_impl->createImage(data, width, height);
 }
 } // karin
